Initial commit of init-unlock
This project is intended to provide a static standalone binary
for acquiring key materiel to be used in decrypting the OS volume.
This inherently means running from the constrained environment of
an init-ramdisk, with no expecation of common userspace being
available.

The first target objective is to support Canonical's MaaS effort
by providing a compatible machanism for working with their FDE
implementation as currently used for Ceph OSDs - placing the disk
key into a Vault for later retrieval by the provisioned Ceph host.

Down the line, the binary will pull keys from other sources, use
TPM for attestation, storage, for request content inputs, and
transform/decrypt acquired key materiel using system data sources.

This commit includes a very simple retrieval process - using the
host's UUID as the Vault token and its serial number to key into
the vault for the value of the crypto keyfile. The initram hook
should require no more than executing the binary to produce the
/crypto_keyfile.bin used by dm-crypt startup hooks when present.

Notes:
  In order to provide a binary which is usable across all Linux
distributions which MaaS may deploy, the build product must be
fully independent of the userspace in which it runs - built on MUSL
libc. Architecture independence is a different matter however, and
ARM64, RISC, and other binary targets will need to be built
separately.
  Building these binaries to be platform/libc independent requires
building OpenSSL statically into the binary, increasing its size
significantly. If the Vault client goes to RustTLS, the resulting
size of this binary should significantly decrease as well.

Testing:
  Only basic bench testing to verify proper data acquisition and
error handling on the client-side. Vault interactions and actual
decrypt using the acquired materiel not yet tested.